feat(src): implements SASS
