+++ netflix/package.json
@@ -19,7 +19,9 @@
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "react-scripts start",
+    "build:css": "node-sass src/ -o src/ ",
+    "start-css": "npm run build:css && node-sass src/ -o src/ -w -r",
+    "start": "npm run build:css && react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
@@ -51,6 +53,7 @@
     "eslint-config-xo-typescript": "^0.55.1",
     "eslint-plugin-react": "^7.31.11",
     "node-sass": "^7.0.3",
+    "npm-run-all": "^4.1.5",
     "typescript": "^4.9.3"
   }
 }
