+++ netflix/src/components/main/films/Film.tsx
@@ -5,7 +5,7 @@ const Film = (props: film) => {
   return (
     <div className="filmContainer">
       <img src={props.img} alt="placeholder" />
-      <div className="details">
+      <div className="--details">
         <h2>{props.name}</h2>
         <h4>{props.year}</h4>
         <h3>{props.genre}</h3>
