+++ netflix/src/components/util/ListItem.tsx
@@ -2,7 +2,7 @@ const ListItem = (props: any) => {
   const value = props.value;
   return (
     // Wrong! There is no need to specify the key here:
-    <li key={value.toString()}>{value}</li>
+    <li>{value}</li>
   );
 };
