+++ netflix/src/index.scss
@@ -1,0 +1,174 @@
+$bgColor: #424242;
+$textColor: #ffffff;
+$accentColor: #f65261;
+$trimColor: #555555;
+$extra-color: #232323;
+$font-Family-1: "Noto Sans", sans-serif;
+$font-family-2: "Source Sans Pro", sans-serif;
+
+body {
+  background-color: $bgColor;
+  overflow-y: scroll;
+}
+
+button {
+  margin-left: 5px;
+  height: 4vh;
+  width: 10%;
+  border-radius: 5px;
+  padding: 5px;
+  text-align: center;
+  position: relative;
+  outline: none;
+  text-decoration: none;
+  border: none;
+  font-size: x-large;
+  &:hover {
+    color: $textColor;
+    border: 1px $textColor;
+    cursor: pointer;
+  }
+}
+.--pri {
+  @extend button;
+  background-color: $accentColor;
+}
+.--sec {
+  @extend button;
+  background-color: transparent;
+}
+
+.App {
+  background-color: $bgColor;
+  min-height: 100%;
+  min-width: 100%;
+  height: 100%;
+  position: absolute;
+  top: 0;
+  left: 0;
+  color: $textColor;
+}
+
+.searchContainer {
+  background-image: url("https://i.picsum.photos/id/681/5000/2951.jpg?hmac=01BxiMvFITUx4h3xM_x0IUp2w5KyQHzUd1AVxm8FKG0");
+  height: 10%;
+  width: 100%;
+  padding: 5vh;
+  padding-top: 7vh;
+  border: 4vh;
+  font-size: 2vh;
+  display: inline-flex;
+  position: relative;
+  background-size: cover;
+}
+.--submit {
+  @extend .--pri;
+  top: 4vh;
+  left: 70vw;
+}
+.--addMovie {
+  @extend .--sec;
+  position: absolute;
+  color: $textColor;
+  top: 1vh;
+  right: 15vh;
+  border: 1px solid $textColor;
+  &:hover {
+    border: 1px solid $accentColor;
+    color: $accentColor
+  }
+}
+
+.searchBar {
+  background: $extra-color;
+  height: 2vh;
+  width: 60%;
+  border-radius: 5px;
+  opacity: 85%;
+  padding: 1vh;
+  text-align: left;
+  position: absolute;
+  top: 10vh;
+  color: $textColor;
+  font-size: x-large;
+  box-shadow: inset 0 0 5px $textColor;
+}
+
+.searchResults {
+  justify-content: left;
+  text-align: left;
+  padding-left: 2vh;
+}
+
+.sortOption {
+  margin-right: 2vw;
+  &:hover {
+    cursor: pointer;
+    box-shadow: 0 5px 15px $textColor;
+    border-radius: 10px;
+  }
+}
+
+.context {
+  position: absolute;
+  height: 4vh;
+  width: 4vh;
+  top: 10px;
+  right: 10px;
+  color: $textColor;
+  &:hover {
+    color: $accentColor
+  }
+}
+
+.film {
+  margin: 2vh;
+  width: fit-content;
+  justify-items: center;
+  position: relative;
+  list-style: none;
+  &:hover {
+    cursor: pointer;
+  }
+}
+
+.mainContainer {
+  background: $extra-color;
+}
+
+.filmsContainer {
+  background: var(--extra-color);
+  display: grid;
+  width: 100%;
+  grid-template-columns: 34% 33% 33%;
+  justify-content: left;
+}
+
+.--details {
+  display: grid;
+  grid-template-columns: 70% auto;
+}
+
+.genreBar {
+  display: flex;
+  justify-content: left;
+  grid-column: span 2;
+  color: var(--text-color);
+}
+.sortBar {
+  display: grid;
+  grid-template-columns: 40% 40% 20%;
+  border-bottom: 4px solid white;
+  padding: 10px;
+  margin: 1rem;
+}
+.sortBy {
+  display: grid;
+  grid-template-columns: auto auto;
+}
+
+.logo {
+  position: absolute;
+  color: $accentColor;
+  top: 0vh;
+}
